requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
gTTS==2.5.4
openai==1.40.6
groq==0.11.0
google-generativeai==0.7.2
python-telegram-bot[rate-limiter]==21.6
requests==2.32.3
wikipedia==1.4.0